[FIX] : DB 데이터 불러오기 방식 변경 및 리사이클러뷰 화면 갱신 수정
