+++ app/src/main/java/com/example/matching_manager/ui/match/MatchRepository.kt
@@ -1,10 +1,6 @@
 package com.example.matching_manager.ui.match
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-
 interface MatchRepository {
-    fun getList(): LiveData<MutableList<MatchDataModel>>
-    fun addData(data : MatchDataModel) {
-    }
+    suspend fun getList(): List<MatchDataModel>
+    suspend fun addData(data: MatchDataModel)
 }
