--- app/src/main/java/com/example/matching_manager/ui/match/MatchRepository2.kt
@@ -1,6 +1,0 @@
-package com.example.matching_manager.ui.match
-
-interface MatchRepository2 {
-    suspend fun getList(): List<MatchDataModel>
-    fun addData(data: MatchDataModel)
-}
