+++ app/src/main/java/com/example/matching_manager/ui/match/MatchRepositoryImpl.kt
@@ -1,46 +1,30 @@
 package com.example.matching_manager.ui.match
 
-import android.service.autofill.FieldClassification.Match
-import android.util.Log
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import com.example.matching_manager.R
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.ValueEventListener
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
+import kotlinx.coroutines.tasks.await
 
-class MatchRepositoryImpl()  : MatchRepository{
+class MatchRepositoryImpl() : MatchRepository {
 
-    val database = Firebase.database("https://matching-manager-default-rtdb.asia-southeast1.firebasedatabase.app/")
+    val database =
+        Firebase.database("https://matching-manager-default-rtdb.asia-southeast1.firebasedatabase.app/")
     val matchRef = database.getReference("Match")
 
-
-    override fun getList(): LiveData<MutableList<MatchDataModel>> {
-            val mutableData = MutableLiveData<MutableList<MatchDataModel>>()
-        matchRef.addValueEventListener(object : ValueEventListener {
-            val listData: MutableList<MatchDataModel> = mutableListOf<MatchDataModel>()
-            override fun onDataChange(snapshot: DataSnapshot) {
-                if (snapshot.exists()){
-                    listData.clear()
-                    for (matchSnapshot in snapshot.children){
-                        val getData = matchSnapshot.getValue(MatchDataModel::class.java)
-                        listData.add(getData!!)
-                        mutableData.value = listData
-                    }
-
-                    Log.d("mutable", "getList() = listData : ${listData.size}")
+    override suspend fun getList(): List<MatchDataModel> {
+        val items = arrayListOf<MatchDataModel>()
+        val snapshot = matchRef.get().await()
+        if (snapshot.exists()) {
+            for (matchSnapshot in snapshot.children) {
+                 matchSnapshot.getValue(MatchDataModel::class.java)?.let { matchData ->
+                     items.add(matchData)
                 }
             }
-            override fun onCancelled(error: DatabaseError) {
-                TODO("Not yet implemented")
-            }
-        })
-        return mutableData
+        }
+
+        return items
     }
 
-    override fun addData(data: MatchDataModel) {
-        matchRef.push().setValue(data)
+    override suspend fun addData(data: MatchDataModel) {
+        matchRef.push().setValue(data).await()
     }
 }
