+++ app/src/main/java/com/example/matching_manager/ui/match/MatchViewModel.kt
@@ -4,24 +4,28 @@ import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 
 class MatchViewModel(private val repository: MatchRepository) : ViewModel() {
 
-    private val _list : MutableLiveData<List<MatchDataModel>> = MutableLiveData()
-    val list : LiveData<List<MatchDataModel>> get() = _list
+    private val _list: MutableLiveData<List<MatchDataModel>> = MutableLiveData()
+    val list: LiveData<List<MatchDataModel>> get() = _list
 
     var matchId = 25
 
-    fun fetchData() {
-        repository.getList().observeForever{
-            val currentList = it
-            Log.d("MatchViewModel", "fetchData() = currentList : ${currentList.size}")
+    suspend fun fetchData() {
 
-            _list.value = currentList
+        val currentList =withContext(Dispatchers.IO) {
+            repository.getList()
         }
+        Log.d("MatchViewModel", "fetchData() = currentList : ${currentList.size}")
+
+        _list.value = currentList
+
     }
 
-    fun addMatch(data : MatchDataModel) {
+    suspend fun addMatch(data: MatchDataModel) {
         repository.addData(data)
     }
 }
